# Optimize casting generic writer to primitives.